+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to
   await jobState.deleteData('abc'); // void
   await jobState.getData('abc'); // undefined
   ```
+### Removed
+
+- Removed `j1-integration compare` command. Developers should use
+  `j1-integration diff` in its place.
 
 ## 6.8.0 - 2021-06-29